renamed file of UIToggle child classes to match classes name (so that usages are shown correctly in Rider)
