+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/AvatarEditorHUD/Scripts/TMPColorOnToggle.cs
@@ -1,13 +1,13 @@
-﻿using UnityEngine;
-using UnityEngine.UI;
+﻿using TMPro;
+using UnityEngine;
 
-public class ImageColorOnToggle : UIToggle
+public class TMPColorOnToggle : UIToggle
 {
-    [SerializeField] private Image targetImage;
+    [SerializeField] private TextMeshProUGUI targetText;
 
     [SerializeField] private Color onColor;
 
     [SerializeField] private Color offColor;
 
-    protected override void OnValueChanged(bool isOn) { targetImage.color = isOn ? onColor : offColor; }
+    protected override void OnValueChanged(bool isOn) { targetText.color = isOn ? onColor : offColor; }
 }
